+++ frontend/lost/src/components/SIA/lost-sia/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Frontend annotation time measurement 
   * Annotation time is now measured in frontend, based on user events. For each annotation individual user interaction time is measured. 
+### Fixed
+- Do not lose polygon annotation when hitting enter in create mode
 
 ## [0.2.2] - 2020-08-06
 ### Fixed