+++ frontend/lost/src/components/SIA/lost-sia/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.1] - 2020-12-16
+## [0.4.0] - 2020-12-16
 ### Added
 - A possible label can now be selected as default label by id
 - Added maxAnnos to canvasConfig. This allows to define a maximum number of annotations that are allowed per image.